Replace int func with float

+++ backend/delivery/views.py
@@ -9,9 +9,10 @@ class RestaurantListView(ListAPIView):
     serializer_class = RestaurantModelSerializer
     
     def get_queryset(self):
-        longitude = int(self.request.query_params.get('long'))
-        latitude = int(self.request.query_params.get('lat'))
+        longitude = float(self.request.query_params.get('long'))
+        latitude = float(self.request.query_params.get('lat'))
         user_location = Point(longitude, latitude, srid=4326)
+        
         queryset = Restaurant.objects.annotate(distance=Distance('location', user_location)).order_by('distance')[0:6]
         
         return queryset
